chore: fix all clippy warnings and remove unused dependencies
- Fixed ~40 clippy warnings across all crates
- Removed unused deps: anyhow, hex, rand, tracing-appender (fortify-core)
- Removed unused deps: hyper-staticfile, tracing-subscriber, urlencoding (fortify-gate)
- Removed unused deps: chrono, tracing-subscriber (fortify-orchestrator)
- Prepares codebase for hyper 1.x migration

+++ fortify/crates/fortify-orchestrator/src/bitmap_stub.rs
@@ -107,7 +107,7 @@ pub fn generate_captcha_image(text: &str) -> Vec<u8> {
     // Build BMP file
     let file_header_size = 14u32;
     let info_header_size = 40u32;
-    let pixel_data_size = (stride * height) as u32;
+    let pixel_data_size = stride * height;
     let file_size = file_header_size + info_header_size + pixel_data_size;
 
     let mut bmp = Vec::with_capacity(file_size as usize);
